requirements.txt
Flask>=3.0
python-docx>=1.1
pdfplumber>=0.11
gunicorn>=22.0
fpdf2>=2.8
requests>=2.31
beautifulsoup4>=4.12
trafilatura>=1.12
openai>=1.30
httpx>=0.27
python-dotenv>=1.0
authlib>=1.3
Flask-SQLAlchemy>=3.1
psycopg2-binary>=2.9
razorpay>=1.4
google-auth>=2.28
google-api-python-client>=2.117
